CPlot: different colormap for each variable
git-svn-id: https://elsa-svn.onera.fr/CASSIOPEE/Trunk/Cassiopee/Apps/Modules@4194 2bbd7fe0-4929-0410-bb6b-e55522daae51

+++ CPlot/CPlot/CPlotState.h
@@ -165,13 +165,13 @@ struct CPlotState {
     int   selectionStyle;  // style pour la selection (0: bleue, 1: alpha)
     int   colormap;        // colormap type
     int   colormapSize;    // number of colors in colormap
-    double* colormapR; // colormap red
-    double* colormapG; // colormap green
-    double* colormapB; // colormap blue
+    float* colormapR; // colormap red
+    float* colormapG; // colormap green
+    float* colormapB; // colormap blue
 
-    double colormapR1, colormapG1, colormapB1; // starting color for bi/tricolor colormaps
-    double colormapR2, colormapG2, colormapB2; // ending color for bi/tricolor colormaps  
-    double colormapR3, colormapG3, colormapB3; // mid color for tricolor colormaps
+    float colormapR1, colormapG1, colormapB1; // starting color for bi/tricolor colormaps
+    float colormapR2, colormapG2, colormapB2; // ending color for bi/tricolor colormaps  
+    float colormapR3, colormapG3, colormapB3; // mid color for tricolor colormaps
 
     int   isoLight;        // light ou pas light -> isoLight
     int   niso;            // nbre d'isos (global)
